+++ core/src/services/metrics.rs
@@ -126,19 +126,15 @@ impl MetricService {
         }
         Ok(())
     }
-    
 
-
-
-
-    fn compute_stats(values: &[f64], is_pose_error: bool) -> StatisticalMetrics {
+    fn compute_stats(values: &[f32], is_pose_error: bool) -> StatisticalMetrics {
         let mut sorted = values.to_vec();
         sorted.sort_by(|a, b| a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal));
 
-        let mean = sorted.iter().sum::<f64>() / sorted.len() as f64;
+        let mean = sorted.iter().sum::<f32>() / sorted.len() as f32;
         let variance = sorted.iter()
             .map(|v| (v - mean).powi(2))
-            .sum::<f64>() / sorted.len() as f64;
+            .sum::<f32>() / sorted.len() as f32;
         let std_dev = variance.sqrt();
 
         StatisticalMetrics {
@@ -147,7 +143,7 @@ impl MetricService {
             min: *sorted.first().unwrap_or(&0.0),
             max: *sorted.last().unwrap_or(&0.0),
             std: std_dev,
-            rmse: is_pose_error.then(|| (values.iter().map(|v| v.powi(2)).sum::<f64>() / values.len() as f64).sqrt()),
+            rmse: is_pose_error.then(|| (values.iter().map(|v| v.powi(2)).sum::<f32>() / values.len() as f32).sqrt()),
             sse: is_pose_error.then(|| values.iter().map(|v| v.powi(2)).sum()),
         }
     }
